pyproject.toml
[project]
name = "linux-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.59.5",
    "python-dotenv>=1.0.1",
]
